Moved the EventArgs types to a sub-folder.
