+++ Com.Okmer.GameController/Components/EventArgs/SpeedChangeArgs.cs
