+++ Com.Okmer.GameController/Components/EventArgs/ThumbChangeArgs.cs
